+++ Network/Telegram/API/Bot/Object/Sender.hs
@@ -1,5 +1,4 @@
-module Network.Telegram.API.Bot.Object.Update.Message.From
-	(From (..), nickname, firstname, lastname) where
+module Network.Telegram.API.Bot.Object.Sender (Sender (..), nickname, firstname, lastname) where
 
 import "aeson" Data.Aeson (FromJSON (parseJSON), Object, withObject, (.:), (.:?))
 import "aeson" Data.Aeson.Types (Parser)
@@ -15,37 +14,37 @@ import "base" Text.Show (Show)
 import "lens" Control.Lens (Lens')
 import "text" Data.Text (Text)
 
-data From
+data Sender
 	= Bot Int (Maybe Text) Text (Maybe Text) (Maybe Text)
 	| User Int (Maybe Text) Text (Maybe Text) (Maybe Text)
 	deriving Show
 
-instance Eq From where
+instance Eq Sender where
 	Bot i _ _ _ _ == Bot i' _ _ _ _ = i == i'
 	User i _ _ _ _ == User i' _ _ _ _ = i == i'
 	_ == _ = False
 
-type Whom = Int -> Maybe Text -> Text -> Maybe Text -> Maybe Text -> From
+type Whom = Int -> Maybe Text -> Text -> Maybe Text -> Maybe Text -> Sender
 
-instance FromJSON From where
-	parseJSON = withObject "From" $ \v -> v .: "is_bot"
-		>>= bool (from v User) (from v Bot) where
+instance FromJSON Sender where
+	parseJSON = withObject "Sender" $ \v -> v .: "is_bot"
+		>>= bool (sender v User) (sender v Bot) where
 
-		from :: Object -> Whom -> Parser From
-		from v f = f <$> v .: "id"
+		sender :: Object -> Whom -> Parser Sender
+		sender v f = f <$> v .: "id"
 			<*> v .:? "username"
 			<*> v .: "first_name"
 			<*> v .:? "last_name"
 			<*> v .:? "language_code"
 
-nickname :: Lens' From (Maybe Text)
+nickname :: Lens' Sender (Maybe Text)
 nickname f (Bot uid nn fn ln lng) = (\nn' -> Bot uid nn' fn ln lng) <$> f nn
 nickname f (User uid nn fn ln lng) = (\nn' -> User uid nn' fn ln lng) <$> f nn
 
-firstname :: Lens' From Text
+firstname :: Lens' Sender Text
 firstname f (Bot uid nn fn ln lng) = (\fn' -> Bot uid nn fn' ln lng) <$> f fn
 firstname f (User uid nn fn ln lng) = (\fn' -> User uid nn fn' ln lng) <$> f fn
 
-lastname :: Lens' From (Maybe Text)
+lastname :: Lens' Sender (Maybe Text)
 lastname f (Bot uid nn fn ln lng) = (\ln' -> Bot uid nn fn ln' lng) <$> f ln
 lastname f (User uid nn fn ln lng) = (\ln' -> User uid nn fn ln' lng) <$> f ln