+++ CHANGELOG.md
@@ -45,3 +45,4 @@
 * Put `Notification` object into `Callback` submodule
 * Put `Keyboard` submodule into `Message` submodule
 * Put `Callback`, `Message` and `Moving` modules into `Update` submodule
+* Rename `From` to `Sender` and move it from `Message` and `Update` submodules